requirements.txt
Django==5.0.6
elasticsearch-dsl==7.4.1
elasticsearch==7.17.0
django-elasticsearch-dsl==7.2.2
daphne==4.1.1
channels==4.2.2
celery==5.5.0
psycopg2-binary==2.9.9
gunicorn==22.0.0
requests==2.31.0
djangorestframework==3.14.0
django-taggit==4.0.0
chardet==5.2.0
django-celery-beat==2.6.0
redis==5.0.1
django-redis==5.4.0